move custom `cp` io functions into a dedicated module
